+++ truck-geotrait/src/traits/surface.rs
@@ -40,8 +40,8 @@ pub trait ParametricSurface: Clone {
     fn try_range_tuple(&self) -> (Option<Tuple>, Option<Tuple>) {
         let ((u0, u1), (v0, v1)) = self.parameter_range();
         (
-            bound2opt(u0).and_then(move |u0| bound2opt(u1).map(move |u1| (u0, u1))),
-            bound2opt(v0).and_then(move |v0| bound2opt(v1).map(move |v1| (v0, v1))),
+            bound2opt(u0).zip(bound2opt(u1)),
+            bound2opt(v0).zip(bound2opt(v1)),
         )
     }
     /// `None` in default; `Some(period)` if periodic w.r.t. parameter u.